+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## UNRELEASED
+
+- Add vector retrieval and search by vector to collections [#343](https://github.com/hypermodeAI/runtime/pull/343)
+
 ## 2024-08-16 - Version 0.11.1
 
 - Improve logger registration code [#335](https://github.com/hypermodeAI/runtime/pull/335)